feat(install): resolve scripts and run them
I've settled on the idea of managing two separate directories. One for
the defaults that are pulled into src/packages from the repo and another
- packages/ - containing user specific modifications. The idea being
that it should be possible to modify the defaults without having to
write another bash script that makes modifications prior to running the
installer.

This commit also introduces the concept of a .ignore containing all the
default packages that the user does not wish to install. Packages listed
inside the .ignore file filter the scripts that will be run. Even the
user specific ones under packages/