feat(install): example installation scripts
While it might initially make sense to create an array with all the
packages to be installed that quickly becomes hard to work with once you
want to make changes to it. Users will have to resort to manually
modifying files prior to install/update OR write elaborate shell scripts
that pull in the data, modify it and overwrite the file's contents. Way
too complicated.

Instead, every package has it's own script and makes use of a couple of
helper methods to install it and make all the necessary modifications to
the system or .settings files in order to work with it. All neatly kept
together in one place. Overriding simply requires you to create a copy
of the script with the same name in the same location under packages/
instead of src/packages and the changes are automatically picked up on.