feat(install): example optional hyprland install
The same concept and logic can be applied to install optional packages
such as hyprland or qtile (omitted here). We might end up with a long
list of shell scripts but the overal logic necessary in order to install
or update the configuration can be kept to a minimum and is quite easy
to grasp. Even for those not too familiar with shell scripting (such as
myself!)